Refactor test utilities
